refactor: 15줄 넘는 함수 분할, render.js 이름 변경
